#9: Add modal for delete confirmation

+++ chronos-frontend/src/app/modules/admin/views/admin-entries/admin-entries.component.ts
@@ -1,9 +1,12 @@
-import {Component, OnInit} from '@angular/core';
+import {Component, OnInit, TemplateRef, ViewChild} from '@angular/core';
 import {AdminEntriesService} from "../../services/admin-entries.service";
 import {Observable, of} from "rxjs";
 import {Entry} from "../../../../model/entry.model";
 import {faPenToSquare, faPlus, faSearch, faTrash} from "@fortawesome/free-solid-svg-icons";
 import {ActivatedRoute, Router} from "@angular/router";
+import {NgbModal} from "@ng-bootstrap/ng-bootstrap";
+import {AdminConfirmModal} from "../../components/admin-confirm-modal/admin-confirm-modal";
+import {AdminConfirmService} from "../../services/admin-confirm.service";
 
 
 @Component({
@@ -21,9 +24,13 @@ export class AdminEntriesComponent implements OnInit {
   searchIcon = faSearch;
   deleteIcon = faTrash;
 
+  @ViewChild('content', {static: false})
+  modalContent!: TemplateRef<void>;
+
   constructor(private adminEntriesService: AdminEntriesService,
               private router: Router,
-              private route: ActivatedRoute) {
+              private route: ActivatedRoute,
+              private confirmService: AdminConfirmService) {
   }
 
   ngOnInit(): void {
@@ -39,9 +46,22 @@ export class AdminEntriesComponent implements OnInit {
   }
 
   deleteEntry(entry: Entry): void {
-    this.adminEntriesService.deleteEntry(entry).subscribe(() => {
-      this.entries$ = this.adminEntriesService.allEntries();
-    })
+    this.confirmService.confirm(
+      "Confirm deleting entry",
+      `Do you want to delete entry ${entry.title}?`
+    )
+      .then(
+      (result) => {
+        console.log("Deleting")
+        this.adminEntriesService.deleteEntry(entry).subscribe(() => {
+          this.entries$ = this.adminEntriesService.allEntries();
+        })
+      },
+      (reason) => {
+        console.log("Leaving")
+        //this.closeResult = `Dismissed ${reason}`;
+      },
+    );
   }
 
   filterEntry(entry: Entry): boolean {
